Move schema module and add submodules (#122)
* Move `schema` module out of internal and make it a direct child of `worker`.
* Move schema primitives into `primitives` submodule.
* Move schema data objects into submodules.
