+++ spatialos-sdk/src/worker/internal/mod.rs
@@ -1,2 +1,1 @@
-pub mod schema;
 pub(crate) mod utils;
