+++ spatialos-sdk/src/worker/mod.rs
@@ -10,6 +10,7 @@ pub mod metrics;
 pub mod op;
 pub mod parameters;
 pub mod query;
+pub mod schema;
 pub mod snapshot;
 pub mod vtable;
